2054: Add detailed CSV export for survey
